ai-service/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
sentence-transformers>=3.0.0
transformers>=4.45.0
torch>=2.2.0
google-genai>=1.69.0
